[CDEC-500] Factor out example datatypes from Json+Bi tests
The `Bi` & `Json` modules had a number of datatypes which were shared
and could be factored out. This commit pulls them out into the
`ExampleHelpers` module.
