+++ chain/test/Test/Pos/Chain/Block/Bi.hs
@@ -26,10 +26,11 @@ import           Pos.Crypto (Hash, SignTag (..), abstractHash, createPsk, hash,
 import           Test.Pos.Binary.Helpers.GoldenRoundTrip (goldenTestBi,
                      roundTripsBiBuildable, roundTripsBiShow)
 import           Test.Pos.Chain.Block.Gen
-import           Test.Pos.Core.Bi (exampleBlockVersion, exampleChainDifficulty,
-                     exampleEpochIndex, exampleLightDlgIndices,
-                     examplePublicKey, exampleSecretKey, exampleSecretKeys,
-                     exampleSlotId, exampleSlotLeaders, exampleSoftwareVersion,
+import           Test.Pos.Core.ExampleHelpers (exampleBlockVersion,
+                     exampleChainDifficulty, exampleEpochIndex,
+                     exampleLightDlgIndices, examplePublicKey,
+                     exampleSecretKey, exampleSecretKeys, exampleSlotId,
+                     exampleSlotLeaders, exampleSoftwareVersion,
                      exampleSscPayload, exampleSscProof, exampleTxPayload,
                      exampleTxProof, exampleUpdatePayload, exampleUpdateProof,
                      feedPM, feedPMC, staticHeavyDlgIndexes,
